trunk/version-1-0/webapp/requirements.txt
Django==1.9.2
argparse==1.4.0
djangorestframework==3.3.2
dnspython==1.12.0
gunicorn==19.4.5
optional-django==0.3.0
pygtail==0.6.1
python-dateutil==2.4.2
react==2.0.2
requests==2.9.1
setuptools==20.1.1
six==1.10.0
wheel==0.29.0
wsgiref==0.1.2
